Refactor some codes such css, api call, util, and set search params operation fuction in vans page
